Programs involving moving balls.
moving_ball.py is a program that creates a ball in a tkinter canvas, with a random velocity. It bounces around the confines of its canvas.

colliding_balls.py is a program that deals with collision detection of a large amount of particles, similarily to moving_ball.py the particles generated in the canvas have random velocities and position, and are free to roam the confines of thier canvas. Written in Python.